Split up model tests a bit more
